Rename school do file
