+++ apps/server/src/shared/domain/domainobject/index.ts
@@ -3,7 +3,7 @@ export * from './pseudonym.do';
 export * from './video-conference.do';
 export * from './board';
 export * from './user-login-migration.do';
-export * from './school.do';
+export * from './legacy-school.do';
 export * from './user.do';
 export * from './page';
 export * from './role-reference';
