+++ apps/server/src/shared/domain/domainobject/legacy-school.do.ts
